Manipulation de la bd v2.0

+++ Program.cs
@@ -45,11 +45,12 @@ Console.WriteLine("Donner votre choix");
 Console.WriteLine("1: Ajouter");
 Console.WriteLine("2: Supprimer");
 Console.WriteLine("3: Mettre à jour");
+Console.WriteLine("4: Afficher les postes d'un utilisateur");
 int x=0;
 do{
 x=Convert.ToInt32(Console.ReadLine());
 }
-while ( x<1 || x >3 );
+while ( x<1 || x >4 );
 User user=new User();
 switch (x) {
     case 1:
@@ -57,23 +58,36 @@ switch (x) {
         string name=Console.ReadLine();
         user.Name=name;        
         new UserRepository().Add(user);  
+        Console.WriteLine("--------------------------");
+        Console.WriteLine("Votre nouvelle Base de données :");
+        new UserRepository().Select();
     break;
     case 2:
         Console.WriteLine("Donner l'id");
         int y =Convert.ToInt32(Console.ReadLine());
         user.Id=y;
         new UserRepository().Delete(y);  
+        Console.WriteLine("--------------------------");
+        Console.WriteLine("Votre nouvelle Base de données :");
+        new UserRepository().Select();
     break;
     case 3:
         Console.WriteLine("Donner l'id");
         int y1 =Convert.ToInt32(Console.ReadLine());
         user.Id=y1;
         new UserRepository().Update(y1);
+        Console.WriteLine("--------------------------");
+        Console.WriteLine("Votre nouvelle Base de données :");
+        new UserRepository().Select();
+    break;
+    case 4:
+        Console.WriteLine("Donner l'id");
+        int y2 =Convert.ToInt32(Console.ReadLine());
+        user.Id=y2;
+        new UserRepository().Postes(y2);
     break;
 }
-Console.WriteLine("--------------------------");
-Console.WriteLine("Votre nouvelle Base de données :");
-new UserRepository().Select();
+
 public class Post
 {
     public int Id { get; set; }
@@ -134,6 +148,7 @@ public class UserRepository
         else
         { 
             connection.Execute("INSERT INTO User (Name) VALUES (@Name)",new{user.Name});    
+            Console.WriteLine("Ajout terminé avec succès");
         }        
     }
 
@@ -141,7 +156,21 @@ public class UserRepository
     {
         var connection = new SqliteConnection("DataSource=./data/demo1.db");
         if(GetById(id)!=null){
-            connection.Execute("DELETE FROM User WHERE Id=@id", new {id});
+            var result=connection.Query<User>("SELECT * FROM POST p , User u WHERE u.Id=p.OwnerId and p.OwnerId=@id",new{id});
+            if(result.Any()){
+                Console.WriteLine("Cet utilisateur a déja " +result.Count()+" poste(s).");
+                Console.WriteLine("Voulez-vous le supprimer ?");
+                string answer="";
+                do{
+                    answer=Console.ReadLine();
+                }while(answer.ToUpper()!="OUI" && answer.ToUpper()!="NON");
+                if(answer.ToUpper()=="YES"){
+                    connection.Execute("DELETE FROM User WHERE Id=@id", new {id});
+                    Console.WriteLine("Suppression terminée avec succès");
+                }else{
+                    Console.WriteLine("Suppression annulée");
+                }
+            }
         }else{
             Console.WriteLine("Utilisateur n'existe pas déja");
         }
@@ -152,17 +181,39 @@ public class UserRepository
         var connection = new SqliteConnection("DataSource=./data/demo1.db");
         if(GetById(id)!=null){
             Console.WriteLine("Donner le nouveau nom");
-        string name=Console.ReadLine();
-        connection.Execute("UPDATE User SET Name=@name WHERE Id=@id", new {id, name});      
+            string name=Console.ReadLine();
+            connection.Execute("UPDATE User SET Name=@name WHERE Id=@id", new {id, name});      
+            Console.WriteLine("Mise à jour terminée avec succès");
         }else{
             Console.WriteLine("Utilisateur n'existe pas");
+            
         }  
     }
     public void Select (){
         var connection = new SqliteConnection("DataSource=./data/demo1.db");
         var result=connection.Query<User>("SELECT * FROM User ");
+        Console.WriteLine("Table User :");
         foreach(var x in result){
             Console.WriteLine(x);
+        }
+    }
+    public void Postes(int id){
+        var connection = new SqliteConnection("DataSource=./data/demo1.db");
+        var result=connection.Query<User>("SELECT * FROM User WHERE Id=@id",new{id});
+        if(result.Any()){
+            var result1=connection.Query<Post>("SELECT * FROM User u,Post p WHERE u.Id=p.OwnerId and p.OwnerId=@id",new{id});
+            if(result1.Any()){
+                Console.WriteLine("Cet utilisateur a "+result1.Count()+" poste(s):");
+                Console.WriteLine("");
+                foreach(var x in result1){
+                    Console.WriteLine("Post "+x.Id+"= Title: "+x.Title+" | Content: "+x.Content);
+                    Console.WriteLine("");
+                }
+            }else{
+                Console.WriteLine("Cet utilisateur n'a aucune poste");
+            }
+        }else{
+            Console.WriteLine("Cet utilisateur n'existe pas");
         }
     }
 }
